Complete 01/07_Camera

+++ src/LIBs/Utility/EventManager/EventManager.cpp
@@ -14,21 +14,21 @@ EventManager * EventManager::GetInstance() {
 
 //------------
 
-void EventManager::RegisterOp(int event, Ptr<Operation> & op) {
+void EventManager::RegisterOp(size_t event, Ptr<Operation> & op) {
 	if (directory.find(event) == directory.end())
 		directory[event] = Ptr<OpQueue>(new OpQueue, Operation::Delete);
 	directory[event]->Push(op);
 }
 
-void EventManager::RegisterOp(int event, Operation * op) {
+void EventManager::RegisterOp(size_t event, Operation * op) {
 	RegisterOp(event, Ptr<Operation>(op, Operation::Delete));
 }
 
-void EventManager::RegisterOp(int event, const std::function<void ()> & op) {
+void EventManager::RegisterOp(size_t event, const std::function<void ()> & op) {
 	RegisterOp(event, new LambdaOp(op));
 }
 
-void EventManager::Response(int event) {
+void EventManager::Response(size_t event) {
 	auto it = directory.find(event);
 	if (it != directory.end())
 		it->second->Run();
